% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{A binomial function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{
Returns a table of the binomial values, and returns a graph illustrating the number of successes of the simulation
}
\description{
A binomial function
}
\examples{
\dontrun{mybin(iter = 10000, n = 20, p = .9)}
}
