% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function that creates confidence intervals to estimate the value of the function applied to a sample.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{How many iterations of the sample.}

\item{x}{The sample that the iterations are done over}

\item{fun}{The function that we will be looking at for our confidence interval}

\item{alpha}{The interval we want to look at, say alpha is .05 then we look at the confidence interval of 95% or 1-alpha.}

\item{cx}{The text size. Default is 50% larger.}

\item{...}{Extra parameters for our histogram to take, like 'xlab', 'ylab', or other hitogram functions.}
}
\value{
Outputs the confidence interval, function, and the sample x to the console, and outputs a histogram that estimates our function through the confidence interval.
}
\description{
myboot2 function that creates confidence intervals to estimate the value of the function applied to a sample.
}
\examples{
\dontrun{sam=rnorm(10,mean=10,sd=5);myboot2(x=sam,iter=10000,fin="medium",alpha=.05,xlab="medium")}
}
