% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Calculates 95% confidence interval for mu of the data}
\usage{
myci(x)
}
\arguments{
\item{x}{The sample data we want to estimate mu for}
}
\value{
A 95% confidence interval.
}
\description{
Calculates 95% confidence interval for mu of the data
}
\examples{
#sam=rnorm(10,mean=10,sd=5);myci(sam)
}
