% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Uniform Distribution approximation}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{The size of each sample space}

\item{iter}{The amount of sample spaces}

\item{a}{Lower bounds of the sample space}

\item{b}{Upper bounds of the sample space}
}
\value{
Outputs a histogram of the sample mean along with a red theoretical uniform curve, a blue density uniform curve, and a black mean curve.
}
\description{
Uniform Distribution approximation
}
\examples{
\dontrun{mycltu(n=10,iter=10000,a=0,b=10)}
}
