% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My maximum likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The function we want to apply to our data.}

\item{x}{The data.}

\item{param}{A sequence of numbers of any length}

\item{...}{Extra parameters for creating a main, x, or y labels, or other things you want to do with plot()}
}
\value{
Returns a graph of the y vs param, and console output
}
\description{
My maximum likelihood
}
