% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution with probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standard deviation}

\item{a}{P(Y<=a)}
}
\value{
A graph with the normal distribution and the area under the curve up to a.
}
\description{
Normal Distribution with probability
}
\examples{
\dontrun{myncurve(10,5,10)}
}
