% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{An introductory plot}
\usage{
myplot(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A plot
}
\description{
An introductory plot
}
\examples{
\dontrun{d <- 1:40; myplot(x = d)}
}
