% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{P value area graph}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{The calculation of our t statistic.}

\item{xmax}{The x range of values that we want to be graphed.}

\item{n}{The size of our sample.}

\item{alpha}{The confidence level 1-alpha.}
}
\value{
Returns the p-value and t statistic, as well as a graph of the region where we reject the null hypothesis indicated by two vertical lines.
}
\description{
P value area graph
}
